Support IE Indeterminate checkboxes for "All" option
IE handles indeterminate checkboxes "differently", it doesn't fire an
onChange event when a checkbox goes from indeterminate to a determinate
state.

This issue is well documented on the jquery github issue:
https://github.com/jquery/jquery/issues/1698

The generally accepted solution is to use the onClick event instead of the
onChange event. You can see this solution implemented in the google's angular
material design library in their github link here:

https://github.com/angular/material2/issues/3145

I applied a similar approach to normalize the events we listen to for the
indeterminate all checkbox and commented why those things are done in the
code so this issue won't regress.