Updated treeview.component.html
Any elements that have the "fa" class have been surrounded by a "span" tag. This is because Font Awesome 5 removes the target element from the DOM and replaces it with an SVG. This means that simply changing the fa-* class will not change the displayed icon. This is compatible with Font Awesome 4 and 5.